Avoid leaking external crates into API
