+++ takumi-core/src/lib.rs
@@ -34,7 +34,6 @@ pub use error::{Error, Result};
 use xxhash_rust::xxh3::Xxh3DefaultBuilder;
 
 pub use crate::resources::font::Fonts;
-pub use parley::Language;
 
 /// Type alias for HashSet using XXH3 hasher
 pub(crate) type Xxh3HashSet<T> = HashSet<T, Xxh3DefaultBuilder>;
