+++ takumi-css/src/style/mod.rs
@@ -4,8 +4,7 @@ mod css_input;
 mod math;
 mod media_query;
 mod properties;
-/// Selectors and stylesheet rule types.
-pub mod selector;
+pub(crate) mod selector;
 mod sizing;
 mod stylesheets;
 mod supports;
@@ -14,15 +13,16 @@ pub mod tw;
 
 pub use animation::apply_stylesheet_animations;
 pub use animation::{KeyframeRule, KeyframesRule};
-pub use calc::{CalcArena, CalcFormula, CalcLinear, parse_calc_number_expression};
+pub(crate) use calc::CalcArena;
+pub(crate) use calc::parse_calc_number_expression;
 pub(crate) use css_input::{CssInput, CssNumber, CssUnexpected, CssValueSeed};
 pub(crate) use math::lerp;
 pub use math::{fast_div_255, fast_div_255_u32};
 pub(crate) use properties::unexpected_token;
 pub use properties::*;
-// Selector matching internals (CssRule, SelectorImpl, Ident, …) stay namespaced
+// Selector matching internals (CssRule, SelectorImpl, Ident, …) stay crate-private
 // under `selector::` for the renderer's `layout::matching`; only the stylesheet
-// entry point is part of the public surface.
-pub use selector::StyleSheet;
+// entry point and `@property` rule type are part of the public surface.
+pub use selector::{MediaQueryList, PropertyRule, StyleSheet};
 pub use sizing::SizingContext;
 pub use stylesheets::*;
