+++ takumi-html/Cargo.toml
@@ -12,6 +12,7 @@ rust-version = "1.91"
 html5ever.workspace = true
 markup5ever_rcdom.workspace = true
 thiserror.workspace = true
+typed-builder.workspace = true
 
 [dependencies.takumi-core]
 path = "../takumi-core"
