+++ takumi-napi/src/measure_task.rs
@@ -1,7 +1,6 @@
 use std::{collections::HashMap, mem::take, sync::Arc};
 
 use napi::bindgen_prelude::*;
-use takumi_core::Language;
 use takumi_core::layout::{DEFAULT_DEVICE_PIXEL_RATIO, Viewport, node::Node, style::StyleSheet};
 use takumi_raster::measure;
 
@@ -21,7 +20,7 @@ pub struct MeasureTask {
   pub(crate) stylesheet: StyleSheet,
   pub(crate) images: HashMap<Arc<str>, (Buffer, ImageCacheMode)>,
   pub(crate) font_families: Option<Vec<String>>,
-  pub(crate) lang: Option<String>,
+  pub(crate) lang: Option<Arc<str>>,
 }
 
 impl MeasureTask {
@@ -60,7 +59,7 @@ impl MeasureTask {
         })
         .collect::<Result<_>>()?,
       font_families: options.font_families,
-      lang: options.lang,
+      lang: options.lang.map(Arc::from),
     })
   }
 }
@@ -86,7 +85,7 @@ impl Task for MeasureTask {
       .node(node)
       .fonts(&fonts)
       .font_families(take(&mut self.font_families))
-      .lang(take(&mut self.lang).and_then(|s| Language::parse(&s).ok()))
+      .lang(take(&mut self.lang))
       .build();
 
     measure(options).map_err(map_error)
