+++ takumi-napi/src/render_task.rs
@@ -1,7 +1,6 @@
 use std::{collections::HashMap, mem::take, sync::Arc};
 
 use napi::bindgen_prelude::*;
-use takumi_core::Language;
 use takumi_core::layout::{DEFAULT_DEVICE_PIXEL_RATIO, Viewport, node::Node, style::StyleSheet};
 use takumi_raster::{DitheringAlgorithm, render, write_image};
 
@@ -26,7 +25,7 @@ pub struct RenderTask {
   pub(crate) stylesheet: StyleSheet,
   pub(crate) images: HashMap<Arc<str>, (Buffer, ImageCacheMode)>,
   pub(crate) font_families: Option<Vec<String>>,
-  pub(crate) lang: Option<String>,
+  pub(crate) lang: Option<Arc<str>>,
 }
 
 impl RenderTask {
@@ -70,7 +69,7 @@ impl RenderTask {
         })
         .collect::<Result<_>>()?,
       font_families: options.font_families,
-      lang: options.lang,
+      lang: options.lang.map(Arc::from),
     })
   }
 }
@@ -98,7 +97,7 @@ impl Task for RenderTask {
         .node(node)
         .fonts(&fonts)
         .font_families(take(&mut self.font_families))
-        .lang(take(&mut self.lang).and_then(|s| Language::parse(&s).ok()))
+        .lang(take(&mut self.lang))
         .draw_debug_border(self.draw_debug_border)
         .build(),
     )
