+++ takumi-napi/src/svg_render_task.rs
@@ -1,7 +1,6 @@
 use std::{collections::HashMap, mem::take, sync::Arc};
 
 use napi::bindgen_prelude::*;
-use takumi_core::Language;
 use takumi_core::layout::{Viewport, node::Node, style::StyleSheet};
 
 use crate::{
@@ -19,7 +18,7 @@ pub struct SvgRenderTask {
   pub(crate) stylesheet: StyleSheet,
   pub(crate) images: HashMap<Arc<str>, (Buffer, ImageCacheMode)>,
   pub(crate) font_families: Option<Vec<String>>,
-  pub(crate) lang: Option<String>,
+  pub(crate) lang: Option<Arc<str>>,
 }
 
 impl SvgRenderTask {
@@ -53,7 +52,7 @@ impl SvgRenderTask {
         })
         .collect::<Result<_>>()?,
       font_families: options.font_families,
-      lang: options.lang,
+      lang: options.lang.map(Arc::from),
     })
   }
 }
@@ -80,7 +79,7 @@ impl Task for SvgRenderTask {
         .node(node)
         .fonts(&fonts)
         .font_families(take(&mut self.font_families))
-        .lang(take(&mut self.lang).and_then(|s| Language::parse(&s).ok()))
+        .lang(take(&mut self.lang))
         .build(),
     )
     .map_err(map_error)
