+++ takumi-wasm/src/export.ts
@@ -1,9 +1,5 @@
 import {
   Renderer as RendererInternal,
-  type ByteBuf,
-  type Font,
-  type FontDetails,
-  type ImageSource,
   type Node,
   type RegisteredFamily,
   type RenderAnimationOptions as RenderAnimationOptionsInternal,
@@ -14,267 +10,125 @@ import {
 export * from "../pkg/takumi_wasm";
 export { default } from "../pkg/takumi_wasm";
 
-import { fontFromUrl, subsetFonts } from "@takumi-rs/helpers";
+import { subsetFonts } from "@takumi-rs/helpers";
+import { FontRegistry } from "@takumi-rs/helpers/renderer";
+import type {
+  AnimationOutputFormatOptions,
+  FontLoader,
+  ImagesInput,
+  OutputFormatOptions,
+} from "@takumi-rs/helpers/renderer";
 
-/**
- * A font to register. Either a URL string (fetched on demand, with name/weight/style read from the
- * file), raw bytes, or a descriptor with a lazy `data()` loader.
- */
-export type FontLoader =
-  | string
-  | Font
-  | (Omit<FontDetails, "data"> & {
-      data: () => Promise<FontDetails["data"]> | FontDetails["data"];
-      /** Inclusive codepoint ranges this face covers; lets `render` skip it when unused. */
-      ranges?: [number, number][];
-    } & ({ key: string } | { name: string }));
+export type {
+  AnimationOutputFormatOptions,
+  FontLoader,
+  ImageLoader,
+  ImagesInput,
+  OutputFormatOptions,
+} from "@takumi-rs/helpers/renderer";
 
-type ImageLoaderData = ImageSource["data"];
-
-export type ImageLoader = Omit<ImageSource, "data"> & {
-  data: ImageLoaderData | (() => ImageLoaderData | Promise<ImageLoaderData>);
-};
-
-/** Images for a render: pre-fetched entries, or a group with a decode-cache default. */
-export type ImagesInput =
-  | ImageLoader[]
-  | {
-      /** Pre-fetched entries, same as the array form. */
-      sources?: ImageLoader[];
-      /** Decode-cache default for every image this render; a source's own `cache` wins. */
-      cache?: NonNullable<ImageLoader["cache"]>;
-    };
-
-/**
- * Output format. Format-specific options live on the variant that supports them,
- * so `quality` cannot be paired with a lossless format. On wasm, WebP is always
- * lossless (lossy WebP is native-only).
- */
-export type OutputFormatOptions =
-  | { format?: "png" }
-  | { format: "jpeg"; quality?: number }
-  | { format: "webp" }
-  | { format: "ico" }
-  | { format: "raw" };
-
-export type RenderOptions = Omit<RenderOptionsInternal, "images" | "format" | "quality"> &
+export type RenderOptions = Omit<
+  RenderOptionsInternal,
+  "images" | "format" | "quality" | "lossless"
+> &
   OutputFormatOptions & {
     fonts?: FontLoader[];
+    signal?: AbortSignal;
     images?: ImagesInput;
   };
 
-/**
- * Animation output format. On wasm, WebP animation is always lossless (lossy
- * WebP is native-only).
- */
-export type AnimationOutputFormatOptions =
-  | { format?: "webp" }
-  | { format: "apng" }
-  | { format: "gif" };
-
-export type RenderAnimationOptions = Omit<RenderAnimationOptionsInternal, "images" | "format"> &
+export type RenderAnimationOptions = Omit<
+  RenderAnimationOptionsInternal,
+  "images" | "format" | "quality" | "lossless"
+> &
   AnimationOutputFormatOptions & {
     fonts?: FontLoader[];
+    signal?: AbortSignal;
     images?: ImagesInput;
   };
 
 export type SvgRenderOptions = Omit<SvgRenderOptionsInternal, "images"> & {
   fonts?: FontLoader[];
+  signal?: AbortSignal;
   images?: ImagesInput;
 };
 
-async function resolveImageLoaders(images: ImagesInput): Promise<ImageSource[]> {
-  const { sources = [], cache } = Array.isArray(images) ? { sources: images } : images;
-  const bySrc = new Map<string, ImageLoader>();
-
-  for (const image of sources) {
-    bySrc.set(image.src, image);
-  }
-
-  return Promise.all(
-    [...bySrc.values()].map(async ({ src, data, cache: own }) => ({
-      src,
-      data: typeof data === "function" ? await data() : data,
-      cache: own ?? cache,
-    })),
-  );
-}
-
 export class Renderer {
-  private fontsByKey = new Map<string, Promise<RegisteredFamily[]>>();
-  private fontsByData = new WeakMap<ByteBuf, Promise<RegisteredFamily[]>>();
   private inner = new RendererInternal();
-
-  private getFont(key: string | ByteBuf) {
-    return typeof key === "string" ? this.fontsByKey.get(key) : this.fontsByData.get(key);
-  }
-
-  private setFont(key: string | ByteBuf, family: Promise<RegisteredFamily[]>) {
-    if (typeof key === "string") {
-      this.fontsByKey.set(key, family);
-    } else {
-      this.fontsByData.set(key, family);
-    }
-  }
-
-  private deleteFont(key: string | ByteBuf) {
-    if (typeof key === "string") {
-      this.fontsByKey.delete(key);
-    } else {
-      this.fontsByData.delete(key);
-    }
-  }
-
-  private async prepareFonts(fonts: FontLoader[] | undefined) {
-    if (!fonts) {
-      return;
-    }
-
-    const families = await Promise.all(fonts.map(this.registerFont.bind(this)));
-
-    return [...new Set(families.flat().map((f) => f.name))];
-  }
+  private fonts = new FontRegistry<RegisteredFamily>((font) => this.inner.registerFont(font));
 
   async render(node: Node, options?: RenderOptions) {
-    const { fonts, fontFamilies, images, ...rest } = options ?? {};
-    const registeredFamilies = await this.prepareFonts(
+    const { fonts, fontFamilies, signal, images, ...rest } = options ?? {};
+    signal?.throwIfAborted();
+    const resolved = await this.fonts.resolveResources(
       fonts && subsetFonts({ fonts, source: node }),
+      images,
+      fontFamilies,
     );
-    const resolvedImages = images ? await resolveImageLoaders(images) : undefined;
+    signal?.throwIfAborted();
 
-    return this.inner.render(node, {
-      ...rest,
-      images: resolvedImages,
-      fontFamilies: fontFamilies ?? registeredFamilies,
-    });
+    return this.inner.render(node, { ...rest, ...resolved });
   }
 
   async renderAsDataUrl(node: Node, options?: RenderOptions) {
-    const { fonts, fontFamilies, images, ...rest } = options ?? {};
-    const registeredFamilies = await this.prepareFonts(
+    const { fonts, fontFamilies, signal, images, ...rest } = options ?? {};
+    signal?.throwIfAborted();
+    const resolved = await this.fonts.resolveResources(
       fonts && subsetFonts({ fonts, source: node }),
+      images,
+      fontFamilies,
     );
-    const resolvedImages = images ? await resolveImageLoaders(images) : undefined;
+    signal?.throwIfAborted();
 
-    return this.inner.renderAsDataUrl(node, {
-      ...rest,
-      images: resolvedImages,
-      fontFamilies: fontFamilies ?? registeredFamilies,
-    });
+    return this.inner.renderAsDataUrl(node, { ...rest, ...resolved });
   }
 
   async renderSvg(node: Node, options?: SvgRenderOptions) {
-    const { fonts, fontFamilies, images, ...rest } = options ?? {};
-    const registeredFamilies = await this.prepareFonts(
+    const { fonts, fontFamilies, signal, images, ...rest } = options ?? {};
+    signal?.throwIfAborted();
+    const resolved = await this.fonts.resolveResources(
       fonts && subsetFonts({ fonts, source: node }),
+      images,
+      fontFamilies,
     );
-    const resolvedImages = images ? await resolveImageLoaders(images) : undefined;
+    signal?.throwIfAborted();
 
-    return this.inner.renderSvg(node, {
-      ...rest,
-      images: resolvedImages,
-      fontFamilies: fontFamilies ?? registeredFamilies,
-    });
+    return this.inner.renderSvg(node, { ...rest, ...resolved });
   }
 
   async measure(node: Node, options?: RenderOptions) {
-    const { fonts, fontFamilies, images, ...rest } = options ?? {};
-    const registeredFamilies = await this.prepareFonts(
+    const { fonts, fontFamilies, signal, images, ...rest } = options ?? {};
+    signal?.throwIfAborted();
+    const resolved = await this.fonts.resolveResources(
       fonts && subsetFonts({ fonts, source: node }),
+      images,
+      fontFamilies,
     );
-    const resolvedImages = images ? await resolveImageLoaders(images) : undefined;
+    signal?.throwIfAborted();
 
-    return this.inner.measure(node, {
-      ...rest,
-      images: resolvedImages,
-      fontFamilies: fontFamilies ?? registeredFamilies,
-    });
+    return this.inner.measure(node, { ...rest, ...resolved });
   }
 
   async renderAnimation(options: RenderAnimationOptions) {
-    const { fonts, fontFamilies, images, ...rest } = options;
+    const { fonts, fontFamilies, signal, images, ...rest } = options;
+    signal?.throwIfAborted();
     const nodes = options.scenes.map((scene) => scene.node);
-    const registeredFamilies = await this.prepareFonts(
+    const resolved = await this.fonts.resolveResources(
       fonts && subsetFonts({ fonts, source: nodes }),
+      images,
+      fontFamilies,
     );
-    const resolvedImages = images ? await resolveImageLoaders(images) : undefined;
+    signal?.throwIfAborted();
 
-    return this.inner.renderAnimation({
-      ...rest,
-      images: resolvedImages,
-      fontFamilies: fontFamilies ?? registeredFamilies,
-    });
+    return this.inner.renderAnimation({ ...rest, ...resolved });
   }
 
+  registerFont(font: FontLoader) {
+    return this.fonts.register(font);
+  }
+
+  /** Releases the underlying wasm renderer's memory. */
   free() {
     this.inner.free();
   }
-
-  async registerFont(font: FontLoader) {
-    const loader = typeof font === "string" ? fontFromUrl(font) : font;
-    const key = createFontKey(loader);
-
-    const cached = this.getFont(key);
-    if (cached) {
-      return cached;
-    }
-
-    const extracted = extractFontBuffer(loader);
-    // Keep the descriptor's name/subsetOf/weight/style; only the data is resolved.
-    const register = (data: ByteBuf) =>
-      this.inner.registerFont(isBuffer(loader) ? data : { ...loader, data });
-
-    if (isBuffer(extracted)) {
-      const binded = register(extracted);
-
-      this.setFont(key, Promise.resolve(binded));
-
-      return binded;
-    }
-
-    const promise = extracted.then(register).catch((error) => {
-      this.deleteFont(key);
-      throw error;
-    });
-
-    this.setFont(key, promise);
-
-    return promise;
-  }
-}
-
-function extractFontBuffer(font: Exclude<FontLoader, string>) {
-  if (isBuffer(font)) {
-    return font;
-  }
-
-  if (typeof font.data !== "function") {
-    return font.data;
-  }
-
-  return font.data();
-}
-
-function createFontKey(font: Exclude<FontLoader, string>) {
-  if (isBuffer(font)) {
-    return font;
-  }
-
-  if ("key" in font && font.key) {
-    return font.key;
-  }
-
-  if (typeof font.data === "function") {
-    return `${font.name ?? ""}:${font.weight ?? ""}:${font.style ?? ""}`;
-  }
-
-  return font.data;
-}
-
-function isBuffer(data: unknown): data is ByteBuf {
-  return (
-    data instanceof Uint8Array ||
-    data instanceof ArrayBuffer ||
-    (typeof Buffer !== "undefined" && Buffer.isBuffer(data))
-  );
 }
